python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aigp"
version = "1.0.0"
description = "AIGP — AI Governance Proof. Open standard for proving your AI agents used the approved policies, prompts, and tools."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "AIGP Community" },
]
keywords = ["aigp", "ai-governance", "opentelemetry", "openlineage", "governance", "audit", "ai-agents", "tracing", "lineage", "merkle-proof"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]

[project.optional-dependencies]
exporters = [
    "opentelemetry-exporter-otlp>=1.20.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["aigp"]

[project.urls]
Homepage = "https://github.com/open-aigp/aigp"
Documentation = "https://github.com/open-aigp/aigp/tree/main/sdks/python"
Repository = "https://github.com/open-aigp/aigp"
Issues = "https://github.com/open-aigp/aigp/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
